Se genera todo con jinja2

+++ ProcessL0.py
@@ -7,16 +7,17 @@ from ProcessBase import ProcessBase
 import os
 import shutil
 import re
+from templates import TemplateHandler
 class ProcessL0(ProcessBase):
     #'./workspaceL0F/inputDir/'
     def __init__(self, workspace_path, temporal_parameterFile=None, parameterFile=None, config_params = None,adq_id = None, path_to_adq = None):
         super().__init__(workspace_path, temporal_parameterFile, parameterFile, config_params,adq_id, path_to_adq)
     
     def ras_files(self):
-        print("El directorio de trabajo es", self.workspace_path)
+ 
         
         # Busca todos los archivos que contienen "VC" en su nombre
-        files_vc = glob.glob(os.path.join(self.workspace_path, '*VC*'))
+        files_vc = glob.glob(os.path.join(self.path_to_adq, '*VC*'))
         
         # Filtra la lista para excluir los archivos que contienen "VC0" en su nombre
         files = [file for file in files_vc if "VC0" not in file]
@@ -25,25 +26,38 @@ class ProcessL0(ProcessBase):
         files.sort()
         
         return files
-    def adec_xeml0f(self):
-        # Obtén las rutas a los directorios y archivos necesarios
-        l0f_template_dir = os.path.join(self.workspace_path, self.config_params.get('L0F_Templates_Dir'))
-        path_workspace_lof = os.path.join(self.workspace_path, self.adq_id, self.config_params.get('workspace_l0f_input'))
+    def get_recent_dttl(self, dttl_files):
+        # Filtrar la lista de archivos para incluir solo los que terminan en 'xemt'
+        xemt_files = [f for f in dttl_files if f.endswith('xemt')]
         
-        ruta_dttl_tamplate =  os.path.join(l0f_template_dir, "inputdttl_template.xml")
+        # Encuentra el archivo con el valor máximo en la posición 11 después de dividir el nombre del archivo
+        dttl_file = max(xemt_files, key=lambda filename: re.split("_", filename)[10])
 
+        return os.path.basename(dttl_file)
 
-        # Procesa cada archivo .xemt en el directorio de adquisición
-        for filename in os.listdir(self.path_to_adq + "/workspaceL0F/inputDir"):
-            if filename.endswith('.xemt'):
-                self._write_templates(filename, ruta_dttl_tamplate, path_workspace_lof)
+        
+    def adec_xeml0f(self,ras_files, dtt_file):
+        # Obtén las rutas a los directorios y archivos necesarios
+        #l0f_template_dir = os.path.join(self.workspace_path, self.config_params.get('L0F_Templates_Dir'))
+        #path_workspace_lof = os.path.join(self.workspace_path, self.adq_id, self.config_params.get('workspace_l0f_input'))
+        lista_vc_xemt = []
+        #ruta_dttl_tamplate =  os.path.join(l0f_template_dir, "inputdttl_template.xml")
+        for filename in ras_files:
+            f_name = os.path.basename(filename)
+            if re.match(r'S1[AB]_OPER_SAR_RAS____ETT_VC[1-9]_', f_name) and filename.endswith('.xemt'):
+                lista_vc_xemt.append(f_name)
+        dir_to_templates = os.path.join(self.workspace_path, "templates","templates_l0")
+        #print("La direccion a los templates es", dir_to_templates)
+        #print("La direccion a los templates es", dir_to_templates)
+        th = TemplateHandler(dir_to_templates)
+        th.render_ras_file(lista_vc_xemt,dtt_file, 'parameterFile.xml','parameterFile.xml')
+
         
 
 
     def _write_templates(self, filename, ruta_dttl_tamplate, path_workspace_l0f):
         if filename.endswith('.xemt'):
             file_name = re.sub(r'^\./', '', filename)   # elimina el ./ que precede al nombre del archivo
-            print(file_name)
 
             # Obtiene solo el nombre del archivo de ruta_dttl_tamplate
             template_file_name = os.path.basename(ruta_dttl_tamplate)
@@ -66,7 +80,7 @@ class ProcessL0(ProcessBase):
             
             # Eliminar el "./" del nombre del archivo
             dttl_name = dttl_name.replace("./", "")
-            
+            print("El nombre de dttl es:, ",dttl_name)
             # Reemplazar el string "INPUTDTTLNAME" en el archivo template
             template_file = os.path.join(workspace_input_dir,"inputdttl_template.xml")
             with open(template_file, "r") as file:
@@ -80,45 +94,46 @@ class ProcessL0(ProcessBase):
         #self._create_input_temporal_params_file(workspace_input_dir)    
     def _create_input_temporal_params_file(self, workspace_input_dir):    
         # Concatenate the content of the template file to temporal_parameterFile.xml
-        temporal_parameter_file = "temporal_parameterFile.xml"
-        template_file = "inputxemt_template.xml"
+        temporal_parameter_file = os.path.join(workspace_input_dir, "temporal_parameterFile.xml")
+        template_file = "inputdttl_template.xml"
 
         # Move and process the remaining files
         files = sorted(filename for filename in os.listdir(workspace_input_dir) if "VC" not in filename)
 
         for filename in files:
             # Preparation of the file name to be processed
-            # if os.path.exists(os.path.join(workspace_input_dir,filename)):
-            #     os.remove(os.path.join(workspace_input_dir,filename))
-            # shutil.move(filename, workspace_input_dir)
-            if filename.endswith(".xemt"):
-                filename = filename.replace("./", "")
-                print(filename)
+            source_path = os.path.join(workspace_input_dir, filename)
+            destination_path = os.path.join(workspace_input_dir, filename)
+            if os.path.exists(source_path):
+                shutil.move(source_path, destination_path)
+                if filename.endswith(".xemt"):
+                    filename = filename.replace("./", "")
 
-                with open(template_file, "r") as file:
-                    template_content = file.read()
+                    with open(workspace_input_dir + "/" +template_file, "r") as file:
+                        template_content = file.read()
 
-                template_content = template_content.replace("INPUTXEMTNAME", filename)
+                    template_content = template_content.replace("INPUTXEMTNAME", filename)
 
-                with open(template_file, "w") as file:
-                    file.write(template_content)
+                    with open(template_file, "w") as file:
+                        file.write(template_content)
 
-                with open(temporal_parameter_file, "a") as parameter_file:
-                    parameter_file.write(template_content)
-                    
-                    
-            self._create_input_params_file(temporal_parameter_file)
-    def _create_input_params_file(self,temporal_parameter_file):    
+                    with open(temporal_parameter_file, "a") as parameter_file:
+                        parameter_file.write(template_content)
+                                
+                self._create_input_params_file(temporal_parameter_file,workspace_input_dir)
+
+    def _create_input_params_file(self,temporal_parameter_file,workspace_input_dir):    
         # Concatenar los archivos parameterFile1.xml, temporal_parameterFile.xml y parameterFile2.xml en parameterFile.xml
         parameter_file = "parameterFile.xml"
-        with open(parameter_file, "a") as file:
-            for xml_file in ["parameterFile1.xml", temporal_parameter_file, "parameterFile2.xml"]:
-                with open(xml_file, "r") as xml:
+        with open(os.path.join(workspace_input_dir,parameter_file), "a") as file:
+            for xml_file in ["parameterFile1.xml", "temporal_parameterFile.xml", "parameterFile2.xml"]:
+                file_path = f"{workspace_input_dir}/{xml_file}"
+                with open(file_path, "r") as xml:
                     file.write(xml.read())
         
         # Eliminar los archivos temporales
         os.remove(temporal_parameter_file)
-        os.remove("parameterFile1.xml")
-        os.remove("parameterFile2.xml")
-        os.remove("inputdttl_template.xml")
-        os.remove("inputxemt_template.xml")
+        os.remove(os.path.join(workspace_input_dir,"parameterFile1.xml"))
+        os.remove(os.path.join(workspace_input_dir,"parameterFile2.xml"))
+        os.remove(os.path.join(workspace_input_dir,"inputdttl_template.xml"))
+        os.remove(os.path.join(workspace_input_dir,"inputxemt_template.xml"))
